Create SourceExportButton component
