+++ app/src/app/routes/Sources/SourceMappings.tsx
@@ -1,14 +1,8 @@
-import React, { useRef } from "react";
+import React from "react";
 
 import { Icon } from "@blueprintjs/core";
 import { IconNames } from "@blueprintjs/icons";
-import {
-  Container,
-  Grid,
-  makeStyles,
-  CircularProgress,
-  Typography,
-} from "@material-ui/core";
+import { Container, Grid, makeStyles, Typography } from "@material-ui/core";
 import { useTranslation } from "react-i18next";
 import { useHistory, useParams } from "react-router";
 
@@ -18,7 +12,8 @@ import MappingsToolbar from "features/Mappings/MappingsToolbar";
 import NavigationBreadcrumbs from "features/NavigationBreadcrumbs/NavigationBreadcrumbs";
 import CredentialEditButton from "features/Sources/CredentialEditButton";
 import SourceDrawer from "features/Sources/SourceDrawer";
-import { useApiSourcesExportRetrieveQuery } from "services/api/generated/api.generated";
+
+import SourceExportButton from "./SourceExportButton";
 
 const useStyles = makeStyles((theme) => ({
   header: {
@@ -43,30 +38,6 @@ const SourceMappings = (): JSX.Element => {
   const history = useHistory();
   const { sourceId } = useParams<{ sourceId?: string }>();
 
-  const exportButtonRef = useRef<HTMLAnchorElement | null>(null);
-  const {
-    data: serializedSource,
-    isLoading: isSerializedSourceLoading,
-  } = useApiSourcesExportRetrieveQuery(
-    { id: sourceId ?? "" },
-    { skip: !sourceId }
-  );
-
-  const handleExportMappingClick = () => {
-    if (serializedSource) {
-      const data = `text/json;charset=utf-8,${encodeURIComponent(
-        JSON.stringify(serializedSource)
-      )}`;
-
-      if (exportButtonRef && exportButtonRef.current) {
-        exportButtonRef.current.setAttribute("href", `data:${data}`);
-        exportButtonRef.current.setAttribute(
-          "download",
-          `${serializedSource.name}.json`
-        );
-      }
-    }
-  };
   const handleLaunchClick = () => {
     history.push(`/sources/${sourceId}/batches`);
   };
@@ -83,20 +54,7 @@ const SourceMappings = (): JSX.Element => {
               color="secondary"
               startIcon={<Icon icon={IconNames.COG} />}
             />
-            <Button
-              component="a"
-              variant="contained"
-              color="secondary"
-              className={classes.button}
-              startIcon={<Icon icon={IconNames.EXPORT} />}
-              ref={exportButtonRef}
-              disabled={isSerializedSourceLoading}
-              onClick={handleExportMappingClick}
-              disableElevation
-            >
-              {isSerializedSourceLoading && <CircularProgress size="small" />}
-              <Typography>{t("exportMapping")}</Typography>
-            </Button>
+            <SourceExportButton />
             <Button
               variant="contained"
               color="secondary"
